+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/TxCert.hs
@@ -74,14 +74,11 @@ module Cardano.Ledger.Shelley.TxCert (
   isUnRegStakeTxCert,
 ) where
 
-import Cardano.Base.Proxy (asProxy)
 import Cardano.Ledger.BaseTypes (invalidKey, kindObjectValue)
 import Cardano.Ledger.Binary (
   DecCBOR (decCBOR),
-  DecCBORGroup (..),
   Decoder,
   EncCBOR (..),
-  EncCBORGroup (..),
   Encoding,
   FromCBOR (..),
   ToCBOR (..),
@@ -91,8 +88,9 @@ import Cardano.Ledger.Binary (
   decodeWord,
   encodeListLen,
   encodeWord8,
-  listLenInt,
+  getDecoderVersion,
   peekTokenType,
+  withCurrentEncodingVersion,
  )
 import Cardano.Ledger.Coin (Coin (..), DeltaCoin)
 import Cardano.Ledger.Core
@@ -105,7 +103,12 @@ import Cardano.Ledger.Internal.Era (AllegraEra, AlonzoEra, BabbageEra, MaryEra)
 import Cardano.Ledger.Keys (asWitness)
 import Cardano.Ledger.Shelley.Era (ShelleyEra)
 import Cardano.Ledger.Shelley.PParams ()
-import Cardano.Ledger.State (StakePoolParams (..))
+import Cardano.Ledger.State (
+  StakePoolParams (..),
+  decodeStakePoolParamsFlat,
+  encodeStakePoolParamsFlat,
+  stakePoolParamsListLen,
+ )
 import Cardano.Ledger.Val ((<+>), (<×>))
 import Control.DeepSeq (NFData (..), rwhnf)
 import Data.Aeson (ToJSON (..), (.=))
@@ -437,9 +440,10 @@ encodeShelleyDelegCert = \case
 encodePoolCert :: PoolCert -> Encoding
 encodePoolCert = \case
   RegPool poolParams ->
-    encodeListLen (1 + listLen (asProxy poolParams))
-      <> encodeWord8 3
-      <> encCBORGroup poolParams
+    withCurrentEncodingVersion $ \v ->
+      encodeListLen (1 + stakePoolParamsListLen v)
+        <> encodeWord8 3
+        <> encodeStakePoolParamsFlat poolParams
   RetirePool vk epoch ->
     encodeListLen 3
       <> encodeWord8 4
@@ -507,8 +511,9 @@ shelleyTxCertDelegDecoder = \case
 poolTxCertDecoder :: EraTxCert era => Word -> Decoder s (Int, TxCert era)
 poolTxCertDecoder = \case
   3 -> do
-    group <- decCBORGroup
-    pure (1 + listLenInt (Just group), RegPoolTxCert group)
+    v <- getDecoderVersion
+    poolParams <- decodeStakePoolParamsFlat
+    pure (1 + fromIntegral (stakePoolParamsListLen v), RegPoolTxCert poolParams)
   4 -> do
     a <- decCBOR
     b <- decCBOR